STRUCTURE: Added 404 not found page #8
