+++ src/pages/404.js
@@ -1,0 +1,11 @@
+import React from "react"
+
+const NotFound = () => {
+  return (
+    <div>
+      <h1>Oops. The page you are looking for is not there</h1>
+    </div>
+  )
+}
+
+export default NotFound
